feat: Modify .header and add padding to .hero
I added position property with fixed attribute for .header but that deformed it so I had to resize it. I then added margin to .hero to push it becuase it was hidden as .header was covering it. Changed some font sizes and font and enclosed everything except .header and .footer inside .content